requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
Django==2.1.8
django-appconf==1.0.3
django-ckeditor==5.7.1
Pillow==6.1.0
PyMySQL==0.9.3
requests==2.22.0
django-hitcount==1.3.1